added classes for error handling and all kinds of updates
